App to save notes
App enables users to create categories of notes initially, then based on the categories, notes can be added to these categories. For each note, audio can be recorded, an image can be captured and attached, and also the location where the note is taken is also recored and all of this is saved as a part of the note.